requirements.txt
# Spyne IDE Requirements
# Production-grade components

# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database (choose based on your database)
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0  # MySQL
# sqlalchemy>=2.0.0  # SQL abstraction (optional)

# LLM Provider
openai>=1.0.0  # OpenAI API

# Authentication
PyJWT>=2.8.0  # JWT tokens

# Caching
redis>=5.0.0  # Redis cache (optional, memory cache available)

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional: Monitoring
# prometheus-client>=0.18.0

# Optional: Structured logging
# structlog>=23.1.0

# Token counting and retry logic
tiktoken>=0.5.0
tenacity>=8.2.0

# Rust Python bindings build tool
maturin>=1.0,<2.0

# HTTP client for NodeRegistry
requests>=2.31.0

# Optional: For async HTTP client
httpx>=0.25.0

# AI SQL System Dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
sqlglot>=24.0.0
networkx>=3.2
pgvector>=0.2.4
llama-index>=0.10.0
trino>=0.330.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sentence-transformers>=2.2.0
torch>=2.0.0

